Comands Goods
Made changes to commands and added more like save, load, and equip also made a folder to store all the values.
Made a Goods folder the interface is unfinised and needs to be worked on
