+++ StarterGame/StarterGame/Commands/LoadCommand.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace StarterGame.Commands
+{
+    public class LoadCommand : Command
+    {
+        public LoadCommand() : base()
+        {
+            this.name = "load";
+        }
+        public override bool execute(Player player)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
