+++ StarterGame/StarterGame/Commands/SaveCommand.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace StarterGame.Commands
+{
+    public class SaveCommand : Command
+    {
+        public SaveCommand() : base()
+        {
+            this.name = "save";
+        }
+        public override bool execute(Player player)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
