+++ StarterGame/StarterGame/Goods/IGoods.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace StarterGame.Goods
+{
+    interface IGoods
+    {
+        
+    }
+}
